SIMparse filecheck
Better logic for the file check and one small bugfix:
a) check each filename argument to see if it has a wildcard, if so, do
a date check. If higher number with older date, give option to include,
exclude or stop. Note: this is fun but it fixes the problem at the
wrong place. I’ll also add a check to SIMrun to make sure it does not
leave those orphan files behind unless expressly told to do so.
b) I changed the regular expression so that it returns the LAST number
in the filename, just in case someone called their simulator output
"2nd_run*.txt"