check for existing files
SIMrun checks for existing output files and asks what to do:
a) delete old
b) rename old
c) abort
New option of overwrite with default false allows silent deletion of
old files without prompt (all files, even if there are more than will
be created now)

Things to consider:
a) do we want to make overwrite=T default?
b) do we want an option to specify new output name at run-time? (I
decided against it and force the user to re-run since I would like to
have the output be consistent with the line used to run it, to avoid
confusion.)
c) is there any utility in allowing the old behavior (i.e. overwrite
but leave stuff behind?) I had the code to rename the old remains but
deleted it since it does not make sense. It should be all or nothing.
Or would anyone want to rewrite simulations 1-5 but keep 6-10?