better version
This may be a better behavior - not bothering you if you just re-run
the same thing, but asking if you changed the number of files. Running
with deleteold=T cleans up first no matter what, avoiding the prompt.

This could be a good balance between 'bother always' and 'never ask'.