a more mature version
Several changes:
a) the function now handles NA values better (catching NA's from vectors that do not enter into equation itself)
b) new samelength() check including warning for NA's, non-numeric and different lengths
c) renamed creat and BW to scr and wt, which seem to be more standard in Pmetrics
d) a bug caught in one of the formulas (double conversion from SI)
e) added another example showing how to easily use the function to add a column to a mdata object
f) renamed the function to PMestGFR to be more consistent with naming
g) added a reference to PMgetCRCL. Should we merge these two? This one can be used on its own or can be cbind-ed to mdata.

I think it is now reasonable mature to be merged into the master.